Removed GIT_CONFIG, git-flow removed readlink -e
GIT_CONFIG: stores the global name, email, and was useful to have it
available within the container.
GIT_CONFIG was attempted to be used to avoid an ln -s to the host's
user's homedir (real ~/.gitconfig). When using this approach, the local
.git/config (local to a given git repo) was not accounted for and
git-flow would report the repo is not initialized.

readlink -e:
This flag doesn't exist in the alpine readlink binary. replaced to -f by
a sed operation now.