__k8s_prompt __aws_prompt addition
Added a __k8s_prompt so that the current used context is visible in the
prompt.
Added a __aws_prompt so that the current account can be seen in the
prompt as well.
They can be enabled and disabled by exporting two variables:
K8S_PS1_SHOW
AWS_PS1_SHOW

Their default is 0 as this is not very tested yet.

Added a scrip to help with the tag and push commands to dockerhub and
quay.io.